+++ src/main/kotlin/kr/hhplus/be/server/repository/product/ProductOptionInventoryRepository.kt
@@ -4,8 +4,8 @@ import kr.hhplus.be.server.domain.product.ProductOptionInventory
 
 interface ProductOptionInventoryRepository {
     fun save(productOptionInventory: ProductOptionInventory): ProductOptionInventory
-    fun findById(inventoryId: Int): ProductOptionInventory?
-    fun findByOptionId(optionId: Int): ProductOptionInventory?
+    fun findById(inventoryId: Long): ProductOptionInventory?
+    fun findByOptionId(optionId: Long): ProductOptionInventory?
     fun update(productOptionInventory: ProductOptionInventory): ProductOptionInventory
-    fun delete(inventoryId: Int)
+    fun delete(inventoryId: Long)
 }
