+++ src/main/kotlin/kr/hhplus/be/server/repository/user/AccountRepository.kt
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.repository.user
 import kr.hhplus.be.server.domain.user.Account
 
 interface AccountRepository {
-    fun findById(accountId: Int): Account?
+    fun findById(accountId: Long): Account?
     fun login(loginId: String, password: String): Account?
     fun save(account: Account): Account
     fun update(account: Account): Account
